Split TransactionForm and TransactionList to components & added useState for message on TransactionForm
