+++ client/src/components/TransactionList/index.js
@@ -1,0 +1,23 @@
+import React from "react";
+
+const TransactionList = () => {
+  return (
+    <div className="col-8 text-center">
+      <div className="col">
+        <div className="m-4">
+          <div id="debit-transaction">
+            You paid
+            <span className="cap"> recipient</span>
+            on date
+          </div>
+          <div id="credit-transaction">
+            <span className="cap">recipient</span>
+            paid you $ credit on date
+          </div>
+        </div>
+      </div>
+    </div>
+  );
+};
+
+export default TransactionList;
